Move classes out of scraper.rb and into lib dir
