+++ lib/members_page.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+require 'scraped'
+
+class MembersPage < Scraped::HTML
+  decorator Scraped::Response::Decorator::AbsoluteUrls
+
+  field :member_urls do
+    noko.css('.item .readmore-link/@href').map(&:text)
+  end
+end
